lab-a2a-agents/pyproject.toml
[project]
name = "lab2-a2a-agents"
version = "0.1.0"
description = "A2A 프로토콜 기반 아키텍처 심의위원회 멀티에이전트 실습"
requires-python = ">=3.11"
dependencies = [
    "python-a2a>=0.5.0",
    "openai>=1.0",
    "python-dotenv>=1.0",
    "flask>=3.0",
    "requests>=2.31",
    "rich>=13.0",
]
